move css in src, remove app directory
